requirements.txt
audiomentations==0.42.0
deepspeed==0.17.1
easydict==1.13
einops==0.8.1
fvcore==0.1.5.post20221221
h5py==3.1.0
librosa==0.11.0
loguru==0.7.3
matplotlib==3.10.3
model_zoo==0.4.1
numpy==1.23.0
Pillow==9.4.0
Pillow==11.3.0
plotly==6.2.0
plotting==0.0.7
prefigure==0.0.10
pytorch_lightning==2.5.1.post0
safetensors==0.5.3
scipy==1.16.0
stable_audio_tools==0.0.19
timm==1.0.17
torch==2.7.1
torchaudio==2.7.1
torchvision==0.22.1
tqdm==4.64.1
